+++ src/pages/PageNotFound.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+const PageNotFound = () => (
+    <section className="PageNotFound">
+        <h3>No results founds</h3>
+    </section>
+)
+
+export default PageNotFound
